added notes directly to recipe card
