+++ MetaCocktailsSwiftData/Model/Notes.swift
@@ -27,7 +27,7 @@ let cloverClubNotes = "The original Clover Club was a gentleman's club in Philad
 
 let harveyWalbangerNotes = "This is a built cocktail that gained popularity through a marketing campaign from George Bednar for McKesson Imports (Galliano's importer at the time). Though, the drink itself was likely the creation of bartender Donato 'Duke' Antone, who may have been serving it since the 1950s at the Hollywood bar Black Watch, calling it the 'Duke Screwdriver'. Bednar picked it up in the late 60s and built an advertising campaign around it for McKesson. It goes to show you that even if the cocktail is total trash, as long as its easy to make and there's tons of advertising behind it, it has a chance to make it into the history books and people like us, who dedicate their entire careers to perfecting their craft, end up writing about it."
 
-let saturnNotes = "FUN FACT: California bartender  J. “Popo” Galsini introduced the Saturn to the world in 1967 when he won the International Bartender’s Association World Championship with it. Originally he titled his gin creation the X-15 after an American rocket plane, but shortly after doing so one of those planes crashed, killing its pilot. Reportedly, Galsini renamed the drink as the Saturn for the competition."
+let saturnNotes = "California bartender  J. “Popo” Galsini introduced the Saturn to the world in 1967 when he won the International Bartender’s Association World Championship with it. Originally he titled his gin creation the X-15 after an American rocket plane, but shortly after doing so one of those planes crashed, killing its pilot. Reportedly, Galsini renamed the drink as the Saturn for the competition."
 
 let aperolSpritzNotes = "Aperol was introduced to the market in 1919, initially enjoyed primarily with soda water. Around the mid-20th century, this bitter liqueur found its way into the classic Italian spritz, enhancing the popular drink with the addition of sparkling wine. A significant turning point came in 2003 when the Campari company acquired Aperol. They launched a highly effective advertising campaign that dramatically increased the drink's global popularity. As a result of this marketing success, the Aperol Spritz has become a universally recognized cocktail."
