+++ MetaCocktailsSwiftData/Views/Recipe View/RecipeBackView.swift
@@ -19,11 +19,7 @@ struct RecipeViewBack: View {
             
             FadingEdgesScrollView {
                 
-                if viewModel.isShowingIngredientRecipe {
-                    IngredientRecipeView(prep: viewModel.currentIngredientRecipe, viewModel: viewModel)
-                        .padding()
-
-                } else if let buildOrder = viewModel.cocktail.buildOrder {
+                if let buildOrder = viewModel.cocktail.buildOrder {
                     BuildOrderView(buildOrder: buildOrder, viewModel: viewModel)
                         .padding()
                 }
@@ -51,8 +47,8 @@ struct BuildOrderView: View {
     
     var body: some View {
         
-        VStack(alignment: .leading) {
-
+        VStack(alignment: .leading, spacing: 12) {
+            
             Grid(horizontalSpacing: 1) {
                 
                 GridRow {
@@ -96,10 +92,10 @@ struct BuildOrderView: View {
                 if build.step > 0 {
                     HStack(alignment: .top) {
                         Text("\(build.step).")
-                            .font(FontFactory.mediumFont(size: 14))
+                            .font(FontFactory.mediumFont(size: 16))
                             .bold()
                         Text("\(build.method)")
-                            .font(FontFactory.mediumFont(size: 14))
+                            .font(FontFactory.mediumFont(size: 16))
                     }
                     .listRowBackground(Color.darkGrey)
                     Divider()
@@ -111,62 +107,3 @@ struct BuildOrderView: View {
     }
 }
 
-struct IngredientRecipeView: View {
-    
-    let prep: Prep
-    var viewModel: RecipeViewModel
-    
-    var body: some View {
-        
-        VStack(alignment: .leading) {
-            
-            Grid(horizontalSpacing: 1) {
-                
-                GridRow {
-                    
-                    Button(action: {
-                        viewModel.flipCard()
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.75) {
-                            viewModel.isShowingIngredientRecipe = false
-                        }
-                    }) {
-                        Image(systemName: "arrow.triangle.2.circlepath")
-                            .foregroundColor(ColorScheme.interactionTint)
-                            .font(.system(size: 22))
-                            .contentShape(Rectangle())
-                    }
-                    .frame(width: 50, height: 50)
-                    .gridColumnAlignment(.leading)
-                    
-                    Text("\(prep.prepIngredientName) recipe")
-                        .font(FontFactory.mediumFont(size: 18).bold())
-                        .gridColumnAlignment(.center)
-                        .frame(maxWidth: .infinity)
-                        .multilineTextAlignment(.center)
-                    
-                    Color.clear
-                        .gridColumnAlignment(.trailing)
-                        .frame(width: 50, height: 50)
-                }
-            }
-            .padding(.bottom, 15)
-            
-            ForEach(prep.prepRecipe) { build in
-                HStack(alignment: .top) {
-                    Text("\(build.step).")
-                        .font(FontFactory.mediumFont(size: 14))
-                        .bold()
-                        
-                    Text("\(build.method)")
-                        .font(FontFactory.mediumFont(size: 14))
-                }
-                .foregroundStyle(.primary)
-                .listRowBackground(Color.darkGrey)
-                
-                Divider()
-            }
-
-            Spacer()
-        }
-    }
-}
