--- MetaCocktailsSwiftData/Views/Recipe View/SubViews/BuildOrderView.swift
@@ -1,16 +1,0 @@
-//
-//  BuildOrderView.swift
-//  MetaCocktails
-//
-//  Created by James on 8/8/23.
-//
-
-import SwiftUI
-
-
-
-//struct BuildOrderView_Previews: PreviewProvider {
-//    static var previews: some View {
-////        BuildOrderView(buildOrder: ramosGinFizz.buildOrder ?? Build(instructions: [Instruction(step: 1, method: "Quit your job")]))
-//    }
-//}
